refactor recurrent + various improvements
hid_init and cell_init accepts TensorVariables. self.hid_init and
self.cell_init are not added as params if TensorVariables are passed in.

hid_out and cell_out are added as self variables to allow them to be
referenced from outside the layer.

Removes a bug when the input dimension is larger than 3

Use the strict=True for scan

Fixed #17

Added tests. The coverage is 97%. The remaining lines are
nonlinearities being set to identity.

Adds grad_clipping as argument. This optionally clips the backward
messages.