Change recurrent layers to accept mask input layer
As discussed here:
https://github.com/Lasagne/Lasagne/pull/294#discussion_r35106579
when mask was supplied as a kwarg to get_output, it's not possible
to pass different masks to different layers in a larger network.  A
good example would be when there are two network branches which are
receiving different input.  We decided here:
https://github.com/Lasagne/Lasagne/issues/337
that the best approach (for now) is to make the recurrent layers
subclass MergeLayer, and have two inputs: One for the actual layer
input, and another for the mask.  This commit implements that change
with a new keyword arg to the constructor of all of the recurrent
layer classes, `mask_input`.  It also updates all of the tests and
the example.