Fix RecurrentLayer type(incoming) == tuple error
RecurrentLayer would throw an error when incoming was a tuple because it
was trying to access incoming.output_shape.  This adds an explicit check
whether incoming is a tuple, and handles it appropriately. Also adds a
test for this special case because all of the RecurrentLayer tests used
a InputLayer as incoming.